requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
reportlab==4.4.3
sqlparse==0.5.3
tzdata==2025.2
